<refactor> Move async function to checkout VM, animate confirm message
