+++ [14] Cupcake Corner/Cupcake Corner/Views/Check Out/CheckoutView.swift
@@ -10,9 +10,6 @@ import SwiftUI
 struct CheckoutView: View {
 	@ObservedObject var viewModel: CheckoutView.ViewModel
 	
-	@State private var confirmationMessage = ""
-	@State private var showingConfirmation = false
-	
 	var body: some View {
 		ScrollView {
 			VStack {
@@ -30,45 +27,22 @@ struct CheckoutView: View {
 				
 				Button("Place Order") {
 					Task {
-						await placeOrder()
+						await viewModel.placeOrder()
 					}
 				}
 				.padding()
+				
+				if viewModel.showingConfirmation {
+					Text(viewModel.confirmationMessage)
+						.bold()
+				}
 			}
+			.animation(.easeInOut, value: viewModel.showingConfirmation)
 		}
 		.navigationTitle("Check out")
 		.navigationBarTitleDisplayMode(.inline)
-		.alert("Thank you!", isPresented: $showingConfirmation) {
-			Button("OK") { }
-		} message: {
-			Text(confirmationMessage)
-		}
 	}
 	
-	func placeOrder() async {
-		// 1. Convert our current order object into some JSON data that can be sent
-		guard let encoded = try? JSONEncoder().encode(self.viewModel.order) else {
-			print("Failed to encode order")
-			return
-		}
-		
-		// 2. Tell Swift how to send that data over a network call
-		let url = URL(string: "https://reqres.in/api/cupcakes")!
-		var request = URLRequest(url: url)
-		request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-		request.httpMethod = "POST"
-		
-		// 3. Run that request and process the response
-		do {
-			let (data, _) = try await URLSession.shared.upload(for: request, from: encoded)
-			// handle the result
-			let decodedOrder = try JSONDecoder().decode(Order.self, from: data)
-			confirmationMessage = "Your order for \(decodedOrder.quantity)x \(Cupcake.types[decodedOrder.item.type].rawValue.lowercased()) cupcakes is on its way!"
-			showingConfirmation = true
-		} catch {
-			print("Checkout failed.")
-		}
-	}
 }
 
 struct CheckoutView_Previews: PreviewProvider {
